fix all the errors (oops)
