+++ src/shared-types.ts
@@ -3,7 +3,7 @@ export interface ServerStartupConfig {
   privateCacheStorage: string;
   apkFilepath: string;
   sdkVersion: number;
-  supportedAbis: string[];
+  supportedAbis: Array<"x86" | "x86_64" | "armeabi-v7a" | "arm64-v8a">;
   version: string;
   buildNumber: string;
   bundleId: string;
